requirements.txt
playwright
google-api-python-client
google-auth-oauthlib
pandas
python-dotenv

